+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Todos los cambios notables en este proyecto serán documentados en este archivo.
 
 El formato está basado en [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.9.0] - 2025-10-19
+
+### Added
+- Implementación de Cifrado Homofónico.
+- Nuevo argumento `genkey homophonic` para generar la clave homofónica.
+
 ## [1.8.0] - 2025-10-19
 
 ### Added